* sysdeps/s390/pt-machine.h (testandset): Use long int as return
value. 
* sysdeps/arm/pt-machine.h (testandset): Likewise. 
* sysdeps/hppa/pt-machine.h (testandset): Likewise. 
* sysdeps/m68k/pt-machine.h (testandset): Likewise. 
* sysdeps/sh/pt-machine.h (testandset): Likewise. 
* sysdeps/sparc/sparc32/pt-machine.h (testandset): Likewise. 
* sysdeps/sparc/sparc64/pt-machine.h (testandset): Likewise.

+++ linuxthreads/sysdeps/hppa/pt-machine.h
@@ -38,7 +38,7 @@ register char * stack_pointer __asm__ ("%r30");
 #define xstr(s) str(s)
 #define str(s) #s
 /* Spinlock implementation; required.  */
-PT_EI int
+PT_EI long int
 testandset (int *spinlock)
 {
   int ret;
